src/telegram/model/requirements.txt
ackermann-msgs==2.0.2
action-msgs==1.2.1
action-tutorials-interfaces==0.20.3
action-tutorials-py==0.20.3
actionlib-msgs==4.2.3
aiohttp==3.9.1
aiosignal==1.3.1
ament-cmake-test==1.3.6
ament-copyright==0.12.9
ament-cppcheck==0.12.9
ament-cpplint==0.12.9
ament-flake8==0.12.9
ament-index-python==1.4.0
ament-lint==0.12.9
ament-lint-cmake==0.12.9
ament-package==0.14.0
ament-pep257==0.12.9
ament-uncrustify==0.12.9
ament-xmllint==0.12.9
angles==1.15.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.2
bond==3.0.2
builtin-interfaces==1.2.1
cachetools==5.3.2
cartographer-ros-msgs==2.0.9000
certifi==2023.11.17
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.20
click==8.1.7
coloredlogs==15.0.1
composition-interfaces==1.2.1
control-msgs==4.4.0
controller-manager==2.35.0
controller-manager-msgs==2.35.0
cv-bridge==3.2.1
dataclasses-json==0.6.3
demo-nodes-py==0.20.3
Deprecated==1.2.14
diagnostic-msgs==4.2.3
diagnostic-updater==3.1.2
distro==1.8.0
domain-coordinator==0.10.0
dwb-msgs==1.1.12
example-interfaces==0.9.3

exceptiongroup==1.2.0
fastapi==0.105.0
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.12.2
gazebo-msgs==3.7.0
generate-parameter-library-py==0.3.6
geometry-msgs==4.2.3
google-auth==2.25.2
googleapis-common-protos==1.62.0
greenlet==3.0.2
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.6
image-geometry==3.2.1
importlib-metadata==6.11.0
importlib-resources==6.1.1
interactive-markers==2.3.2
Jinja2==3.1.2
joblib==1.3.2
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
jsonpatch==1.33
jsonpointer==2.4
kubernetes==28.1.0
langchain==0.0.350
langchain-community==0.0.3
langchain-core==0.1.1
langsmith==0.0.71
laser-geometry==2.4.0
launch==1.0.4
launch-ros==0.19.6
launch-testing==1.0.4
launch-testing-ros==0.19.6
launch-xml==1.0.4
launch-yaml==1.0.4
lifecycle-msgs==1.2.1
logging-demo==0.20.3
map-msgs==2.1.0
MarkupSafe==2.1.3
marshmallow==3.20.1
message-filters==4.3.3
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nav-2d-msgs==1.1.12
nav-msgs==4.2.3
nav2-common==1.1.12
nav2-msgs==1.1.12
nav2-simple-commander==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.5.0
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
osrf-pycommon==2.0.2
overrides==7.4.0
packaging==23.2
pcl-msgs==1.0.0
pendulum-msgs==0.20.3
Pillow==10.1.0
posthog==3.1.0
protobuf==4.25.1
pulsar-client==3.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
PyPika==0.48.9
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-qt-binding==1.1.2
python-telegram-bot==20.7
PyYAML==6.0.1
qt-dotgraph==2.2.2
qt-gui==2.2.2
qt-gui-cpp==2.2.2
qt-gui-py-common==2.2.2
quality-of-service-demo-py==0.20.3
rcl-interfaces==1.2.1
rclpy==3.3.11
rcutils==5.1.4
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
resource-retriever==3.1.1
rmw-dds-common==1.6.0
ros2action==0.18.7
ros2bag==0.15.8
ros2cli==0.18.7
ros2component==0.18.7
ros2controlcli==2.35.0
ros2doctor==0.18.7
ros2interface==0.18.7
ros2launch==0.19.6
ros2lifecycle==0.18.7
ros2multicast==0.18.7
ros2node==0.18.7
ros2param==0.18.7
ros2pkg==0.18.7
ros2run==0.18.7
ros2service==0.18.7
ros2topic==0.18.7
rosapi==1.3.2
rosapi-msgs==1.3.2
rosbag2-interfaces==0.15.8
rosbag2-py==0.15.8
rosbridge-library==1.3.2
rosbridge-msgs==1.3.2
rosbridge-server==1.3.2
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpyutils==0.2.1

rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
scripts==3.7.0
sensor-msgs==4.2.3
sensor-msgs-py==4.2.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
shape-msgs==4.2.3
six==1.16.0
slam-toolbox==2.6.6
smclib==3.0.2
sniffio==1.3.0
SQLAlchemy==2.0.23
sros2==0.10.4
starlette==0.27.0
statistics-msgs==1.2.1
std-msgs==4.2.3
std-srvs==4.2.3
stereo-msgs==4.2.3
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.15.0
topic-monitor==0.20.3
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
trajectory-msgs==4.2.3
transformers==4.36.2
triton==2.1.0
typeguard==4.1.5
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
unique-identifier-msgs==2.2.1
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
visualization-msgs==4.2.3
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
xacro==2.0.8
yarl==1.9.4
zipp==3.17.0
